pyproject.toml
# pyproject.toml

[project]
name = "smart-meeting-agent"
version = "0.1.0"
description = "An AI agent to find upcoming meetings and profile attendees."
requires-python = ">=3.13"
dependencies = ["portia-sdk-python[tools-browser-browserbase]>=0.6.2,<0.7.0",  "python-dotenv>=1.1.1",
  "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "requests",
    "python-dateutil",]
 